+++ src/msl/io/tables.py
@@ -159,9 +159,10 @@ def read_table_excel(  # noqa: PLR0913
         file: The file to read.
         cells: The cells to read. For example, `C9` (i.e, specifying only the top-left cell
             of the table) will start at cell C9 and include all columns to the right and
-            all rows below C9, `A:C` includes all rows in columns A, B and C, and, `C9:G20`
-            includes only the specified cells. If not specified, assumes that the table
-            starts at cell `A1` and returns all cells from the specified `sheet`.
+            all rows below C9, `A:C` includes all rows in columns A, B and C, `C9:G20`
+            includes only the specified cells, and `A:C,E` includes all values in columns
+            A, B, C and E. If not specified, assumes that the table starts at cell `A1`
+            and returns all cells from the specified `sheet`.
         sheet: The name of the sheet to read the data from. If there is only one sheet
             in the workbook then you do not need to specify the name of the sheet.
         as_datetime: Whether dates should be returned as [datetime][datetime.datetime] or
@@ -181,11 +182,7 @@ def read_table_excel(  # noqa: PLR0913
     file = os.fsdecode(file) if isinstance(file, (bytes, str, os.PathLike)) else str(file.name)
 
     with ExcelReader(file, **kwargs) as excel:
-        if cells is not None and not _spreadsheet_range_regex.match(cells):
-            match = _spreadsheet_top_left_regex.match(cells)
-            if not match:
-                msg = f"Invalid cell {cells!r}"
-                raise ValueError(msg)
+        if cells is not None and _spreadsheet_top_left_regex.match(cells):
             name = sheet or excel.sheet_names()[0]
             num_rows, num_cols = excel.dimensions(name)
             letters = excel.to_letters(num_cols - 1)